+++ lib/src/mqtt_client.dart
@@ -249,7 +249,7 @@ class MqttClient {
   Stream<List<MqttReceivedMessage<MqttMessage>>> get updates =>
       subscriptionsManager!.subscriptionNotifier;
 
-  /// Comon client connection method.
+  /// Common client connection method.
   Future<MqttConnectionStatus?> connect(
       [String? username, String? password]) async {
     // Protect against an incorrect instantiation
@@ -271,6 +271,8 @@ class MqttClient {
     connectionHandler.onAutoReconnectRetry = onAutoReconnectRetry;
     connectionHandler.onAutoReconnected = onAutoReconnected;
     connectionHandler.onFailedConnectionAttempt = onFailedConnectionAttempt;
+    connectionHandler.registerForMessage(
+        MqttMessageType.disconnect, _processReceivedDisconnectMessage);
     publishingManager =
         MqttPublishingManager(connectionHandler, clientEventBus);
     authenticationManager ??= MqttAuthenticationManager();
@@ -390,7 +392,7 @@ class MqttClient {
   }
 
   /// Publishes a message to the message broker.
-  /// Returns the message identifer assigned to the message.
+  /// Returns the message identifier assigned to the message.
   /// Raises InvalidTopicException if the topic supplied violates the
   /// MQTT topic format rules.
   int publishMessage(
@@ -516,15 +518,20 @@ class MqttClient {
     }
   }
 
-  /// Actual disconnect processing
-  void _disconnect({bool unsolicited = true}) {
+  // Actual disconnect processing
+  void _disconnect({bool unsolicited = true, fromBroker = false}) {
     // Only disconnect the connection handler if the request is
-    // solicited, unsolicited requests, ie broker termination don't
+    // solicited by the user or the broker, unsolicited requests, i.e. network termination don't
     // need this.
     var disconnectOrigin = MqttDisconnectionOrigin.unsolicited;
     if (!unsolicited) {
-      connectionHandler?.disconnect(disconnectMessage);
-      disconnectOrigin = MqttDisconnectionOrigin.solicited;
+      // Don't send a disconnect message if the disconnect is from the broker.
+      if (!fromBroker) {
+        connectionHandler?.disconnect(disconnectMessage);
+        disconnectOrigin = MqttDisconnectionOrigin.solicited;
+      } else {
+        disconnectOrigin = MqttDisconnectionOrigin.brokerSolicited;
+      }
     }
     publishingManager?.published.close();
     publishingManager = null;
@@ -574,5 +581,16 @@ class MqttClient {
     if (on) {
       MqttLogger.loggingOn = true;
     }
+  }
+
+  // Process a disconnect message received from the broker.
+  bool _processReceivedDisconnectMessage(MqttMessage msg) {
+    final disconnectMsg = msg as MqttDisconnectMessage;
+    final reason = mqttDisconnectReasonCode.asString(disconnectMsg.reasonCode);
+    MqttLogger.log(
+        'MqttClient::_processReceivedDisconnectMessage - Disconnect Message received, reason is \'$reason\'  - disconnecting');
+    _connectionStatus.disconnectMessage = disconnectMsg;
+    _disconnect(unsolicited: false, fromBroker: true);
+    return true;
   }
 }